+++ t/1144-isf-public-tested-by-metadata-audit.t
@@ -236,6 +236,7 @@ my $expected_tested_by = [
     't/1342-isf-transaction-port-actor-constant-widths.t',
     't/1343-isf-data-op-static-width-sources.t',
     't/1344-isf-assemble-static-part-widths.t',
+    't/1345-isf-actor-param-actor-constants.t',
 ];
 
 subtest 'direct ISF tested_by metadata is exact and valid' => sub {